+++ app/Http/Controllers/ESBTPPlanningGeneralController.php
@@ -33,7 +33,7 @@ class ESBTPPlanningGeneralController extends Controller
     protected $planningFilterCatalog;
 
     /** Types LMD exclus du planning général BTS (Licence/Master/Doctorat). */
-    private const LMD_TYPES = ['Licence', 'Master', 'Doctorat'];
+    private const LMD_TYPES = \App\Models\ESBTPNiveauEtude::CYCLES_LMD;
 
     public function __construct(
         PlanningConfigurationService $planningConfigService,